make attributeset pub crate only
