--- opentelemetry-sdk/src/attributes/mod.rs
@@ -1,3 +1,0 @@
-mod set;
-
-pub use set::AttributeSet;
